% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_designs.R
\name{draw_designs}
\alias{draw_designs}
\title{Draw several group-sequential design on one plot, for easy comparison}
\usage{
draw_designs(d_list, d_spec, graph_params = list(z_scale = TRUE, z_limits =
  c(-4, 4), n_x_points = 400, hazard_scale = FALSE), include_delay = FALSE)
}
\description{
\code{draw_designs} calls \code{get_design_points} for a list of group-sequential designs and creates a ggplot2 object. 
Some aspects of the plot can be controlled with \code{graph_params}.
}
\details{
\param{d_list} A list of group-sequential design objects. Usually these are created from \code{gsDesign}.
\param{d_spec} Design specification that is common to all designs being considered. This is list containing the sample size 
for a fixed design, allocation ratio, length of delay, etc. 
\param{graph_params} A list of graphical parameters to control the plot: \code{z_scale} (default = \code{TRUE}, if false will plot
on the treatment effect scale), \code{z_limits}, \code{n_x_points}, \code{hazard_scale} (default = \code{FALSE}, if true will overide 
\code{z_scale} and plot on hazard ratio scale).
\param{include_delay} Should the interim sample sizes take account of the delay between time of recruitment and 
time of outcome.
\return A ggplot2 object.
}

