% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design_points.R
\name{get_design_points}
\alias{get_design_points}
\title{Get the coordintes to plot a picture of a group-sequential design}
\usage{
get_design_points(d, z_limits, n_x_points, include_delay = FALSE,
  d_spec = NULL)
}
\description{
\code{get_design_points} Provides a list of co-ordinates that can be used to visualize a group-sequential design.
}
\details{
\param{d} A group sequential design object. Usually this is created from \code{gsDesign}.
\param{z_limits} The plot limits on the z-scale. A vector of length 2. 
Make sure that \code{z_limits[1]} is less than the lower stopping boundary, and that \code{z_limits[2]} is more than
the upper stopping boundary.
\param{n_x_points} The number of points to include on the x-axis.
\param{include_delay} Should the interim sample sizes take account of the delay between time of recruitment and 
time of outcome.
\param{d_spec} Design specification that is common to all designs being considered. This is list containing the sample size 
for a fixed design, allocation ratio, length of delay, etc. 
\return A list containing x cooridates, y coordinates, and which analysis they apply to.
}

