% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e_info.R
\name{get_e_info}
\alias{get_e_info}
\title{Calculate the expected information fraction of a group-sequential design}
\usage{
get_e_info(d, ez, include_delay = FALSE, d_spec = NULL)
}
\description{
\code{get_e_info} calculates the expected information (relative to a fixed design) of a group-sequential design under a range of effect sizes.
}
\details{
\param{d} A group sequential design object. Usually this is created from \code{gsDesign}.
\param{ez} A vector of expected z values. Note this corresponds to the expected z value from a "power-matched" fixed sample 
design.
\param{include_delay} Should the interim sample sizes take account of the delay between time of recruitment and 
time of outcome.
\param{d_spec} Design specification that is common to all designs being considered. This is list containing the sample size 
for a fixed design, allocation ratio, length of delay, etc. 
\return A vector. The expected information fraction (relative to a fixed design) at each value of \code{ez}.
}

