% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pipeline.R
\name{get_pipeline}
\alias{get_pipeline}
\title{Get the sample size at each interim analysis, with and without delay.}
\usage{
get_pipeline(d, d_spec)
}
\description{
\code{get_pipeline} calculates how many patients have been recruited at each interim analysis, and how many patients have
completed.
\param{d} A group sequential design object. Usually this is created from \code{gsDesign}.
\param{d_spec} Design specification that is common to all designs being considered. This is list containing the sample size 
for a fixed design, allocation ratio, length of delay, etc. 
\return A list containing the number of patients completed, the number of patients in the pipeline, the number of patients
recruited and the fraction of patients recruited (relative to a fixed sample size).
}

