% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power.R
\name{get_power}
\alias{get_power}
\title{Find the power of a group-sequential design}
\usage{
get_power(d, ez)
}
\description{
\code{get_power} calculates the power of a group-sequential design for expected z values (of the power-matched fixed design
equal to \code{ez}).
}
\details{
\param{d} A group sequential design object. Usually this is created from \code{gsDesign}.
\param{ez} A vector of expected z values. Note this corresponds to the expected z value from a "power-matched" fixed sample 
design. 
\return A vector. The power at each value of \code{ez}.
}

